refactor: getAllBookedReservations 테마 서비스 경로로 이동
