+++ src/main/java/roomescape/controller/ReservationController.java
@@ -7,10 +7,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.BookedReservationTimeResponseDto;
 import roomescape.dto.ReservationRequestDto;
 import roomescape.dto.ReservationResponseDto;
 import roomescape.service.ReservationService;
@@ -27,12 +25,6 @@ public class ReservationController {
     @GetMapping("/reservations")
     public List<ReservationResponseDto> readAllReservations() {
         return reservationService.getAllReservations();
-    }
-
-    @GetMapping("/themes/{themeId}/times")
-    public List<BookedReservationTimeResponseDto> readBookedReservationTimes(
-        @RequestParam String date, @PathVariable Long themeId) {
-        return reservationService.getAllBookedReservationTimes(date, themeId);
     }
 
     @PostMapping("/reservations")
